+++ front/src/common/redux/actions/index.js
@@ -1,11 +1,12 @@
 import calculate from "./exampleAction";
 import applicationLoad from "./example";
 import { loginUserAction } from "./authActions";
-import { studentRegistered } from './studentRegistrationAction';
+import { studentRegistered, hideStudentRegisteredModal } from './studentRegistrationAction';
 
 export default {
     applicationLoad,
     calculate,
     loginUserAction,
-    studentRegistered
+    studentRegistered,
+    hideStudentRegisteredModal,
 };
