+++ front/src/common/redux/actions/types.js
@@ -6,7 +6,6 @@ export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS';
 export const LOGIN_USER_ERROR = 'LOGIN_USER_ERROR';
 
 export const REGISTER_STUDENT = 'REGISTER_STUDENT';
-export const VALIDATE_FIRST_NAME_ENG = 'VALIDATE_FIRST_NAME_ENG';
-export const NOT_UNIQUE_EMAIL = 'NOT_UNIQUE_EMAIL';
-export const GROUP_NOT_FOUND = 'GROUP_NOT_FOUND';
-export const REGISTRATION_STUDENT_SUCCESS = 'REGISTRATION_STUDENT_SUCCESS';
+export const REGISTRATION_STUDENT = 'REGISTRATION_STUDENT';
+export const REGISTRATION_STUDENT_ERROR = 'REGISTRATION_STUDENT_ERROR';
+export const HIDE_STUDENT_REGISTERED_MODAL = 'HIDE_STUDENT_REGISTERED_MODAL';
