+++ front/src/components/app/App.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import ExampleForRedux from "../exampleForRedux/ExampleForRedux";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import LoginPage from "../login/LoginPage";
-import StudentRegistrationPage from '../student-registration-page';
+import StudentRegistrationPage from '../student-registration/student-registration-page';
 
 class App extends Component {
     render() {
