+++ front/src/components/student-registration/student-registration-form/index.js
@@ -1,0 +1,3 @@
+import StudentRegistrationForm from './student-registration-form';
+
+export default StudentRegistrationForm;
