+++ front/src/components/student-registration/student-registration-page/index.js
